Add files via upload
This is clone project of Myntra, an online fashion and lifestyle store. In this project I have tried to make a look like a clone of Myntra Website. The project is built using HTML, CSS and JavaScript. I have learned a lot of new things with working on this project. Myntra aims to providing a hassle-free and enjoyable shopping experience to shoppers across the country with the widest range of brands and products on its portal.